+++ Adafruit_TMP007.h
@@ -54,7 +54,7 @@
 class Adafruit_TMP007  {
  public:
   Adafruit_TMP007(uint8_t addr = TMP007_I2CADDR);
-  boolean begin(uint8_t samplerate = TMP007_CFG_4SAMPLE);  // by default go to once per second
+  boolean begin(uint16_t samplerate = TMP007_CFG_4SAMPLE);  // by default go to once per second
 
   int16_t readRawDieTemperature(void);
   int16_t readRawVoltage(void);
